requirements.txt
anyio==3.6.2
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
fastapi==0.92.0
filelock==3.9.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
OpenAIAuth==0.3.2
pycryptodomex==3.17
pydantic==1.10.5
regex==2022.10.31
requests==2.28.2
revChatGPT==2.3.4
rfc3986==1.5.0
sniffio==1.3.0
starlette==0.25.0
tiktoken==0.2.0
typing_extensions==4.5.0
urllib3==1.26.14

python-dotenv~=0.21.1
uvicorn~=0.20.0